# api/openapi.yaml
openapi: 3.1.0
info:
  title: NoteFoundry API
  version: 0.1.0
  description: REST interface for the single-owner NoteFoundry learning workspace.
servers:
  - url: /
paths:
...
  /auth/github/callback:
    get:
      operationId: completeGitHubLogin
      summary: Complete GitHub OAuth authentication
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Owner session created; redirect to the management application.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: GitHub authentication failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ownerSession:
      type: apiKey
      in: cookie
      name: notefoundry_session
  responses:
    InvalidRequest:
      description: Request data is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: The authenticated identity is not authorized.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many authentication requests were sent in the current window.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...